requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.0
asn1crypto==1.5.1
atomicwrites==1.4.0
attrs==21.4.0
blis==0.7.7
catalogue==2.0.7
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cryptography==37.0.2
spacy==3.2.4
cssselect2==0.6.0
cymem==2.0.6
distlib==0.3.4
Django==3.2
django-advanced-filters==1.4.0
django-braces==1.15.0
django-cors-headers==3.11.0
django-filter==21.1
django-multiselectfield==0.1.12
django-transaction-hooks==0.3
django-transactional-cleanup==0.1.15
django-utils-six==2.0
djangorestframework==3.13.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
filelock==3.4.2
future==0.18.2
googletrans==3.0.0
greenlet==1.1.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
html5lib==1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
ibm-cloud-sdk-core==3.15.1
ibm-watson==6.0.0
idna==2.10
iniconfig==1.1.1
Jinja2==3.0.3
langcodes==3.3.0
langdetect==1.0.7
lxml==4.8.0
MarkupSafe==2.0.1
mssql-django==1.1.2
murmurhash==1.0.6
mysql-connector-python==8.0.28
numpy==1.22.2
opencv-python==4.5.5.62
oscrypto==1.3.0
packaging==21.3
pandas==1.4.1
pathy==0.6.1
Pillow==9.1.1
platformdirs==2.4.1
pluggy==1.0.0
preshed==3.0.6
psycopg2==2.9.3
py==1.11.0
pycparser==2.21
pydantic==1.8.2
pyHanko==0.13.1
pyhanko-certvalidator==0.19.5
PyJWT==2.3.0
pymssql==2.2.4
pyodbc==4.0.32
pyparsing==3.0.8
PyPDF3==1.0.6
pytest==7.1.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.6
python-highcharts==0.4.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.9
requests==2.27.1
rfc3986==1.5.0
simplejson==3.17.6
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
spacy-langdetect==0.1.2
spacy-legacy==3.0.9
spacy-loggers==1.0.2
SQLAlchemy==1.4.31
sqlparse==0.4.2
srsly==2.4.2
svglib==1.3.0
thinc==8.0.15
tinycss2==1.1.1
tomli==2.0.1
tqdm==4.64.0
typer==0.4.1
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.2
Unipath==1.1
uritools==4.0.0
urllib3==1.26.9
virtualenv==20.13.1
wasabi==0.9.1
webencodings==0.5.1
websocket-client==1.1.0
whitenoise==6.0.0
xhtml2pdf==0.2.7
xlwt==1.3.0
